Refactors UI by extracting components and enabling package import optimization
Extracts key UI elements into reusable components for better maintainability and readability. Enables experimental package import optimization for performance improvements.
