+++ src/app/page.tsx
@@ -4,7 +4,9 @@ import React from 'react';
 import { FaJava, FaPhp, FaCss3Alt, FaNodeJs, FaHtml5 } from 'react-icons/fa';
 import { SiCplusplus, SiC, SiLua, SiMysql, SiJavascript } from 'react-icons/si';
 import Image from 'next/image';
-import { motion, AnimatePresence } from 'framer-motion';
+import MorphingTitle from './components/MorphingTitle';
+import TechBadge from './components/TechBadge';
+import AnimatedContainer from './components/AnimatedContainer';
 
 const technologies = [
   { name: 'Lua', icon: SiLua },
@@ -19,94 +21,6 @@ const technologies = [
   { name: 'MySQL', icon: SiMysql },
 ];
 
-interface MorphingTitleProps {
-  titles: string[];
-  fadeDuration?: number;
-  cycleDuration?: number;
-  className?: string;
-  style?: React.CSSProperties;
-}
-
-function MorphingTitle({
-  titles,
-  fadeDuration = 300,
-  cycleDuration = 2500,
-  className = '',
-  style = {},
-}: MorphingTitleProps) {
-  const [current, setCurrent] = React.useState(0);
-
-  React.useEffect(() => {
-    let interval: number | null = null;
-    let hideTime: number | null = null;
-    let paused = false;
-    const startCycle = () => {
-      if (interval === null) {
-        interval = window.setInterval(() => {
-          setCurrent((prev) => (prev + 1) % titles.length);
-        }, cycleDuration);
-      }
-    };
-    const stopCycle = () => {
-      if (interval !== null) {
-        clearInterval(interval);
-        interval = null;
-      }
-    };
-
-    const handleVisibilityChange = () => {
-      if (document.hidden) {
-        stopCycle();
-        paused = true;
-        hideTime = Date.now();
-      } else {
-        if (paused) {
-          const elapsed = hideTime ? Date.now() - hideTime : 0;
-          if (elapsed >= fadeDuration) {
-            setCurrent((prev) => (prev + 1) % titles.length);
-          }
-        }
-        paused = false;
-        startCycle();
-      }
-    };
-
-    document.addEventListener('visibilitychange', handleVisibilityChange);
-    startCycle();
-
-    return () => {
-      document.removeEventListener('visibilitychange', handleVisibilityChange);
-      stopCycle();
-    };
-  }, [titles.length, cycleDuration]);
-
-  return (
-    <span
-      className={className}
-      style={{
-        display: 'inline-block',
-        overflow: 'hidden',
-        position: 'relative',
-        minHeight: 24,
-        ...style,
-      }}
-    >
-      <AnimatePresence mode="wait">
-        <motion.span
-          key={current}
-          initial={{ clipPath: 'inset(0 100% 0 0)', opacity: 0 }}
-          animate={{ clipPath: 'inset(0 0% 0 0)', opacity: 1 }}
-          exit={{ clipPath: 'inset(0 100% 0 0)', opacity: 1 }}
-          transition={{ duration: fadeDuration / 1000, ease: [0.6, -0.05, 0.01, 0.99] }}
-          style={{ display: 'inline-block', position: 'relative' }}
-        >
-          {titles[current]}
-        </motion.span>
-      </AnimatePresence>
-    </span>
-  );
-}
-
 export default function Home() {
   const titles = [
     'Backend Developer',
@@ -118,12 +32,7 @@ export default function Home() {
 
   return (
     <>
-      <motion.div
-        initial={{ opacity: 0 }}
-        animate={{ opacity: 1 }}
-        exit={{ opacity: 0 }}
-        transition={{ duration: 0.5 }}
-      >
+      <AnimatedContainer>
         <section className="flex items-center justify-center min-h-screen px-4">
           <div className="bg-neutral-900 border border-neutral-800 p-10 rounded-lg shadow-xl max-w-2xl w-full">
             <div className="flex items-center justify-between gap-6 mb-6">
@@ -160,23 +69,14 @@ export default function Home() {
                 Languages & Frameworks
               </h2>
               <div className="flex flex-wrap gap-2 text-sm">
-                {technologies.map((tech) => {
-                  const Icon = tech.icon;
-                  return (
-                    <span
-                      key={tech.name}
-                      className="cursor-pointer select-none flex items-center gap-2 bg-neutral-800 px-3 py-1 rounded"
-                    >
-                      <Icon className="text-lg text-white" />
-                      <span>{tech.name}</span>
-                    </span>
-                  );
-                })}
+                {technologies.map((tech) => (
+                  <TechBadge key={tech.name} name={tech.name} icon={tech.icon} />
+                ))}
               </div>
             </div>
           </div>
         </section>
-      </motion.div>
+      </AnimatedContainer>
     </>
   );
 }
